+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+- :feature:`-` Add :func:`structlog.ReturnLogger.failure` and :func:`structlog.PrintLogger.failure` as preparation for the new Twisted logging system.
 - :release:`15.2.0 <2015-06-10>`
 - :bug:`- major` Allow empty lists of processors.
   This is a valid use case since `#26 <https://github.com/hynek/structlog/issues/26>`_ has been merged.